Refactored challenge methods into separate files.
